backend/pyproject.toml
[project]
name = "neighbourhood-score"
version = "2.0.0"
description = "Data-driven neighborhood scoring for Bangalore home buyers"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.10.0",
    "pandas>=2.2.0",
    "geopandas>=1.0.0",
    "osmnx>=2.0.0",
    "overpy>=0.7",
    "geopy>=2.4.0",
    "httpx>=0.28.0",
    "shapely>=2.0.0",
    "scikit-learn>=1.6.0",
    "numpy>=2.0.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.42.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.11.0",
    "ty>=0.0.24",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.ty.environment]
python-version = "3.12"

[tool.ty.rules]
# Anthropic SDK returns union types; code already uses `type: ignore` inline
unresolved-attribute = "warn"
no-matching-overload = "warn"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
